requirements.txt
telethon
python-telegram-bot==20.6
